backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.0.4
celery[redis]==5.3.6
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
faiss-cpu==1.8.0
numpy==1.26.4
sentence-transformers==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
networkx==3.3
python-louvain==0.16
lxml==5.2.1
feedparser==6.0.11
beautifulsoup4==4.12.3
tenacity==8.3.0
structlog==24.2.0
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.20.0
aiofiles==23.2.1
tiktoken==0.7.0
rank-bm25==0.2.2
